Removing UFR Class input coords, added vc valid
Tested and found that the QRect start coords passed into the UFR class objects dont do anything, as the Layout moves them, so removed that. Added basic validation for the vc file. Made stylesheet paths properties instead of local vars. Removed the extraneous 'with' statements when applying stylesheets in dropEvent() and removeBtnClicked() event.